Create a common place for saving/reading plugin state (e.g. current reviewer pull request, etc...)
