+++ src/main/kotlin/com/duxtinto/intellij/plugin/github/codereviews/di/impl/dagger/ProjectContainer.kt
@@ -1,7 +1,7 @@
 package com.duxtinto.intellij.plugin.github.codereviews.di.impl.dagger
 
 import com.duxtinto.intellij.plugin.github.codereviews.di.impl.dagger.components.UiInjectors
-import com.duxtinto.intellij.plugin.github.codereviews.di.impl.dagger.modules.ProjectModule
+import com.duxtinto.intellij.plugin.github.codereviews.di.impl.dagger.modules.project.ProjectModule
 import com.duxtinto.intellij.plugin.github.codereviews.di.scopes.ProjectScoped
 import com.intellij.dvcs.repo.VcsRepositoryMappingListener
 import dagger.Subcomponent
