--- src/main/kotlin/com/duxtinto/intellij/plugin/github/codereviews/di/impl/dagger/modules/idea/IdeaFileAndDocumentModule.kt
@@ -1,6 +1,0 @@
-package com.duxtinto.intellij.plugin.github.codereviews.di.impl.dagger.modules.idea
-
-import dagger.Module
-
-@Module
-class IdeaFileAndDocumentModule
