+++ src/main/kotlin/com/duxtinto/intellij/plugin/github/codereviews/di/impl/dagger/modules/project/UserModule.kt
@@ -1,6 +1,7 @@
-package com.duxtinto.intellij.plugin.github.codereviews.di.impl.dagger.modules
+package com.duxtinto.intellij.plugin.github.codereviews.di.impl.dagger.modules.project
 
 import com.duxtinto.intellij.plugin.github.codereviews.data.DataContract
+import com.duxtinto.intellij.plugin.github.codereviews.di.scopes.ProjectScoped
 import com.duxtinto.intellij.plugin.github.codereviews.domain.DomainDataMapper
 import com.duxtinto.intellij.plugin.github.codereviews.domain.user.UserEntity
 import com.duxtinto.intellij.plugin.github.codereviews.net.users.apiv3.UserFetcher
@@ -12,8 +13,10 @@ import dagger.Module
 @Module
 abstract class UserModule {
     @Binds
+    @ProjectScoped
     abstract fun provideFetcher(fetcher: UserFetcher): DataContract.User.Fetcher
 
     @Binds
+    @ProjectScoped
     abstract fun provideResponseMapper(mapper: UserMapper): DomainDataMapper<UserEntity, UserResponse>
 }
