+++ src/main/kotlin/com/duxtinto/intellij/plugin/github/codereviews/di/impl/dagger/modules/project/idea/actions/IdeaActionModule.kt
@@ -1,5 +1,6 @@
-package com.duxtinto.intellij.plugin.github.codereviews.di.impl.dagger.modules.idea.actions
+package com.duxtinto.intellij.plugin.github.codereviews.di.impl.dagger.modules.project.idea.actions
 
+import com.duxtinto.intellij.plugin.github.codereviews.di.scopes.ProjectScoped
 import com.intellij.openapi.actionSystem.ActionManager
 import dagger.Module
 import dagger.Provides
@@ -10,6 +11,7 @@ import dagger.Provides
 ])
 class IdeaActionModule {
     @Provides
+    @ProjectScoped
     fun provideActionManager(): ActionManager {
         return ActionManager.getInstance()
     }
