+++ src/main/kotlin/com/duxtinto/intellij/plugin/github/codereviews/di/impl/dagger/modules/project/idea/actions/IdeaReviewerActionModule.kt
@@ -1,30 +1,21 @@
-package com.duxtinto.intellij.plugin.github.codereviews.di.impl.dagger.modules.idea.actions
+package com.duxtinto.intellij.plugin.github.codereviews.di.impl.dagger.modules.project.idea.actions
 
 import com.duxtinto.intellij.plugin.github.codereviews.di.qualifiers.Reviewee
-import com.duxtinto.intellij.plugin.github.codereviews.domain.pullrequests.PullRequestsByRepositoryInteractor
-import com.duxtinto.intellij.plugin.github.codereviews.domain.repositories.FindGithubRepoForRootFolderInteractor
+import com.duxtinto.intellij.plugin.github.codereviews.di.qualifiers.Reviewer
+import com.duxtinto.intellij.plugin.github.codereviews.di.scopes.ProjectScoped
 import com.duxtinto.intellij.plugin.github.codereviews.ide.acl.presentation.toolwindow.toolbar.RevieweeToolbarFactory
 import com.duxtinto.intellij.plugin.github.codereviews.ide.acl.presentation.toolwindow.toolbar.reviewee.RefreshPullRequestListAction
 import com.duxtinto.intellij.plugin.github.codereviews.presentation.Toolbar
-import com.duxtinto.intellij.plugin.github.codereviews.presentation.reviewee.RevieweeContent
 import com.intellij.openapi.actionSystem.ActionManager
 import com.intellij.openapi.actionSystem.AnAction
 import dagger.Module
 import dagger.Provides
 
 @Module
-class IdeaRevieweeActionModule {
+class IdeaReviewerActionModule {
     @Provides
-    fun provideRefreshPullRequestListAction(
-            findRepoInteractor: FindGithubRepoForRootFolderInteractor,
-            @Reviewee getPullRequestsInteractor: PullRequestsByRepositoryInteractor,
-            presenter: RevieweeContent.Presenter
-    ): RefreshPullRequestListAction {
-        return RefreshPullRequestListAction(findRepoInteractor, getPullRequestsInteractor, presenter)
-    }
-
-    @Provides
-    @Reviewee
+    @Reviewer
+    @ProjectScoped
     fun provideRevieweeToolbarActions(
             refreshPullRequestsAction: RefreshPullRequestListAction
     ): List<AnAction> {
@@ -34,7 +25,8 @@ class IdeaRevieweeActionModule {
     }
 
     @Provides
-    @Reviewee
+    @Reviewer
+    @ProjectScoped
     fun provideToolbarFactory(
             actionManager: ActionManager,
             @Reviewee actions: @JvmSuppressWildcards List<AnAction>
@@ -43,7 +35,8 @@ class IdeaRevieweeActionModule {
     }
 
     @Provides
-    @Reviewee
+    @Reviewer
+    @ProjectScoped
     fun provideToolbar(@Reviewee toolbar: Toolbar.Factory): Toolbar {
         return toolbar.create()
     }
