+++ src/main/kotlin/com/duxtinto/intellij/plugin/github/codereviews/di/impl/dagger/modules/project/pullrequest/PullRequestModule.kt
@@ -1,4 +1,4 @@
-package com.duxtinto.intellij.plugin.github.codereviews.di.impl.dagger.modules.pullrequest
+package com.duxtinto.intellij.plugin.github.codereviews.di.impl.dagger.modules.project.pullrequest
 
 import com.duxtinto.intellij.plugin.github.codereviews.data.pullrequests.PullRequestRepository
 import com.duxtinto.intellij.plugin.github.codereviews.di.qualifiers.Reviewee
@@ -11,12 +11,37 @@ import com.duxtinto.intellij.plugin.github.codereviews.domain.pullrequests.revie
 import com.duxtinto.intellij.plugin.github.codereviews.net.pullrequests.apiv3.PullRequestFetcher
 import com.duxtinto.intellij.plugin.github.codereviews.net.pullrequests.apiv3.PullRequestResponse
 import com.duxtinto.intellij.plugin.github.codereviews.net.pullrequests.apiv3.PullRequestResponseMapper
+import com.duxtinto.intellij.plugin.github.codereviews.presentation.pullrequestlist.columns.ColumnInfoFactory
 import com.duxtinto.intellij.plugin.github.codereviews.services.pullrequests.GithubDescriptionParser
+import com.intellij.util.ui.ColumnInfo
 import dagger.Binds
 import dagger.Module
+import dagger.Provides
 
-@Module
-abstract class CommonModule {
+@Module(includes = [
+    RevieweeModule::class,
+    ReviewerModule::class
+])
+abstract class PullRequestModule {
+    @Module
+    companion object {
+        @Provides
+        @ProjectScoped
+        @Reviewee
+        @JvmStatic
+        fun provideRevieweePullRequestColumnInfoDefaultArray(@Reviewee columnInfoFactory: ColumnInfoFactory): Array<ColumnInfo<PullRequestEntity, *>> {
+            return columnInfoFactory.createDefaultColumns()
+        }
+
+        @Provides
+        @ProjectScoped
+        @Reviewer
+        @JvmStatic
+        fun provideReviewerPullRequestColumnInfoDefaultArray(@Reviewer columnInfoFactory: ColumnInfoFactory): Array<ColumnInfo<PullRequestEntity, *>> {
+            return columnInfoFactory.createDefaultColumns()
+        }
+    }
+
     @Binds
     @Reviewee
     @ProjectScoped
@@ -32,14 +57,18 @@ abstract class CommonModule {
     abstract fun provideGetAllReviewsForInteractor(interactor: GetAllReviewsForInteractor): CodeReviewsByPullRequestInteractor
 
     @Binds
+    @ProjectScoped
     abstract fun providePullRequestDomainContractFetcher(fetcher: PullRequestFetcher): PullRequestDomainContract.Fetcher
 
     @Binds
+    @ProjectScoped
     abstract fun providePullRequestDomainContractRepository(repository: PullRequestRepository): PullRequestDomainContract.Repository
 
     @Binds
+    @ProjectScoped
     abstract fun providePullRequestDomainContractDescriptionParser(parser: GithubDescriptionParser): PullRequestDomainContract.DescriptionParser
 
     @Binds
+    @ProjectScoped
     abstract fun providePullRequestResponseMapper(mapper: PullRequestResponseMapper): DomainDataMapper<PullRequestEntity, PullRequestResponse>
 }
