+++ src/main/kotlin/com/duxtinto/intellij/plugin/github/codereviews/ide/acl/services/github/pullrequests/IdeaPullRequestLoader.kt
@@ -16,7 +16,7 @@ class IdeaPullRequestLoader
     @Inject
     constructor(
         private val connection: GithubConnectionExt,
-        private val pullRequestMapper: DomainDataMapper<PullRequestEntity, PullRequestResponse>)
+        private val pullRequestMapper: @JvmSuppressWildcards DomainDataMapper<PullRequestEntity, PullRequestResponse>)
     : NetContract.PullRequest.Loader {
 
     companion object {
