+++ src/main/kotlin/com/duxtinto/intellij/plugin/github/codereviews/state/PluginStateHandler.kt
@@ -1,0 +1,7 @@
+package com.duxtinto.intellij.plugin.github.codereviews.state
+
+import com.duxtinto.intellij.plugin.github.codereviews.domain.pullrequests.PullRequestEntity
+
+interface PluginStateHandler {
+    var activeReviewerPullRequest: PullRequestEntity?
+}
